BREAKING: unify targets, flatten folder structure
