+++ src/index.ts
@@ -1,5 +1,6 @@
-export * from './shared'
+export * from './Canvas'
 export * from './constants'
+export * from './events'
 export * from './hooks'
 export * from './reconciler'
 export * from './renderer'
