--- src/native/index.ts
@@ -1,5 +1,0 @@
-import '@expo/browser-polyfill'
-
-export * from './Canvas'
-export * from './events'
-export * from '../'
