+++ src/utils.ts
@@ -1,7 +1,19 @@
+import * as React from 'react'
 // @ts-ignore
 import * as OGL from 'ogl'
 import { COLORS, POINTER_EVENTS, RESERVED_PROPS } from './constants'
-import { Instance, InstanceProps } from './types'
+import { useIsomorphicLayoutEffect } from './hooks'
+import { createRoot } from './renderer'
+import {
+  EventHandlers,
+  Instance,
+  InstanceProps,
+  ObjectMap,
+  RenderProps,
+  RootState,
+  SetBlock,
+  Subscription,
+} from './types'
 
 /**
  * Converts camelCase primitives to PascalCase.
@@ -134,47 +146,223 @@ export const applyProps = (instance: Instance, newProps: InstanceProps, oldProps
 }
 
 /**
- * Shallow checks objects.
+ * Collects nodes and programs from a Mesh.
  */
-export const checkShallow = (a: any, b: any) => {
-  // If comparing arrays, shallow compare
-  if (Array.isArray(a)) {
-    // Check if types match
-    if (!Array.isArray(b)) return false
+export const buildGraph = (object: OGL.Mesh) => {
+  const data: ObjectMap = { nodes: {}, programs: {} }
 
-    // Shallow compare for match
-    if (a == b) return true
+  if (object) {
+    object.traverse((obj: any) => {
+      if (obj.name) {
+        data.nodes[obj.name] = obj
+      }
 
-    // Sort through keys
-    if (a.every((v, i) => v === b[i])) return true
+      if (obj.program?.gltfMaterial && !data.programs[obj.program.gltfMaterial.name]) {
+        data.programs[obj.program.gltfMaterial.name] = obj.program
+      }
+    })
   }
 
-  // Atomically compare
-  if (a === b) return true
-
-  return false
+  return data
 }
 
 /**
- * Prepares a set of changes to be applied to the instance.
+ * Creates event handlers, returning an event handler method.
  */
-export const diffProps = (instance: Instance, newProps: InstanceProps, oldProps: InstanceProps = {}) => {
-  // Prune reserved props
-  newProps = filterKeys(newProps, true, ...RESERVED_PROPS)
-  oldProps = filterKeys(oldProps, true, ...RESERVED_PROPS)
+export const createEvents = (state: RootState) => {
+  const handleEvent = (event: PointerEvent, type: keyof EventHandlers) => {
+    // Convert mouse coordinates
+    state.mouse.x = (event.offsetX / state.renderer.width) * 2 - 1
+    state.mouse.y = -(event.offsetY / state.renderer.height) * 2 + 1
 
-  const changedProps: InstanceProps = {}
+    // Filter to interactive meshes
+    const interactive: OGL.Mesh[] = []
+    state.scene.traverse((node: OGL.Mesh) => {
+      // Mesh has registered events and a defined volume
+      if (node.__handlers && node.geometry?.attributes?.position) interactive.push(node)
+    })
 
-  // Sort through props
-  Object.entries(newProps).forEach(([key, value]) => {
-    // Skip primitives
-    if (instance.isPrimitive && key === 'object') return
-    // Skip if props match
-    if (checkShallow(value, oldProps[key])) return
+    // Get elements that intersect with our pointer
+    state.raycaster.castMouse(state.camera, state.mouse)
+    const intersects: Instance[] = state.raycaster.intersectMeshes(interactive)
 
-    // Props changed, add them
-    changedProps[key] = value
-  })
+    // Used to discern between generic events and custom hover events.
+    // We hijack the pointermove event to handle hover state
+    const isHoverEvent = type === 'onPointerMove'
 
-  return changedProps
+    // Trigger events for hovered elements
+    intersects.forEach((object) => {
+      const handlers = object.__handlers
+
+      // Bail if object doesn't have handlers (managed externally)
+      if (!handlers) return
+
+      if (isHoverEvent && !state.hovered.get(object.id)) {
+        // Mark object as hovered and fire its hover events
+        state.hovered.set(object.id, object)
+
+        // Fire hover events
+        handlers.onPointerMove?.({ event, hit: object.hit })
+        handlers.onPointerOver?.({ event, hit: object.hit })
+      } else {
+        // Otherwise, fire its generic event
+        handlers[type]?.({ event, hit: object.hit })
+      }
+    })
+
+    // Cleanup stale hover events
+    if (isHoverEvent || type === 'onPointerDown') {
+      state.hovered.forEach((object: Instance) => {
+        const handlers = object.__handlers
+
+        if (!intersects.length || !intersects.find((i) => i === object)) {
+          // Reset hover state
+          state.hovered.delete(object.id)
+
+          // Fire unhover event
+          if (handlers?.onPointerOut) handlers.onPointerOut({ event, hit: object.hit })
+        }
+      })
+    }
+
+    return intersects
+  }
+
+  return { handleEvent }
+}
+
+/**
+ * Configures rendering internals akin to R3F.
+ */
+export const createInternals = (canvas: HTMLCanvasElement, props: RenderProps): RootState => {
+  // Create or accept renderer, apply props
+  const renderer =
+    props.renderer instanceof OGL.Renderer
+      ? props.renderer
+      : typeof props.renderer === 'function'
+      ? props.renderer(canvas)
+      : new OGL.Renderer({
+          antialias: true,
+          powerPreference: 'high-performance',
+          ...(props.renderer as any),
+          canvas: canvas,
+        })
+  if (props.renderer && typeof props.renderer !== 'function') applyProps(renderer, props.renderer as InstanceProps)
+  const gl = renderer.gl
+  gl.clearColor(1, 1, 1, 0)
+
+  // Flush frame for native
+  if (renderer.gl.endFrameEXP) {
+    const renderFrame = renderer.render.bind(renderer)
+    renderer.render = ({ scene, camera }) => {
+      renderFrame({ scene, camera })
+      renderer.gl.endFrameEXP()
+    }
+  }
+
+  // Create or accept camera, apply props
+  const camera =
+    props.camera instanceof OGL.Camera
+      ? props.camera
+      : new OGL.Camera(gl, { fov: 75, near: 1, far: 1000, ...(props.camera as any) })
+  camera.position.z = 5
+  if (props.camera) applyProps(camera, props.camera as InstanceProps)
+
+  // Create scene
+  const scene = new OGL.Transform(gl)
+
+  // Init rendering internals for useFrame, keep track of subscriptions
+  let priority = 0
+  const subscribed = []
+
+  // Subscribe/unsubscribe elements to the render loop
+  const subscribe = (refCallback: React.MutableRefObject<Subscription>, renderPriority?: number) => {
+    // Subscribe callback
+    subscribed.push(refCallback)
+
+    // Enable manual rendering if renderPriority is positive
+    if (renderPriority) priority += 1
+  }
+
+  const unsubscribe = (refCallback: React.MutableRefObject<Subscription>, renderPriority?: number) => {
+    // Unsubscribe callback
+    const index = subscribed.indexOf(refCallback)
+
+    if (index !== -1) subscribed.splice(index, 0)
+
+    // Disable manual rendering if renderPriority is positive
+    if (renderPriority) priority -= 1
+  }
+
+  // Init event state
+  const mouse = new OGL.Vec2()
+  const raycaster = new OGL.Raycast(gl)
+  const hovered = new Map()
+
+  // Set initial state
+  const state: RootState = {
+    ...props,
+    renderer,
+    gl,
+    camera,
+    scene,
+    priority,
+    subscribed,
+    subscribe,
+    unsubscribe,
+    mouse,
+    raycaster,
+    hovered,
+  }
+
+  // Init root
+  const root = createRoot(canvas, state)
+
+  // Handle callback
+  if (props.onCreated) props.onCreated(state)
+
+  // Animate
+  const animate = (time?: number) => {
+    // Cancel animation if frameloop is set, otherwise keep looping
+    if (props.frameloop === 'never') return cancelAnimationFrame(state.animation)
+    state.animation = requestAnimationFrame(animate)
+
+    // Call subscribed elements
+    subscribed.forEach((ref) => ref.current?.(state, time))
+
+    // If rendering manually, skip render
+    if (priority) return
+
+    // Render to screen
+    renderer.render({ scene, camera })
+  }
+  if (props.frameloop !== 'never') animate()
+
+  return { ...state, root }
+}
+
+/**
+ * Used to block rendering via its `set` prop. Useful for suspenseful effects.
+ */
+export const Block = ({ set }: { set: React.Dispatch<React.SetStateAction<SetBlock>> }) => {
+  useIsomorphicLayoutEffect(() => {
+    set(new Promise(() => null))
+    return () => set(false)
+  }, [])
+
+  return null
+}
+
+/**
+ * Generic error boundary. Calls its `set` prop on error.
+ */
+export class ErrorBoundary extends React.Component<{ set: React.Dispatch<any> }, { error: boolean }> {
+  state = { error: false }
+  static getDerivedStateFromError = () => ({ error: true })
+  componentDidCatch(error: any) {
+    this.props.set(error)
+  }
+  render() {
+    return this.state.error ? null : this.props.children
+  }
 }
