--- src/web/index.ts
@@ -1,3 +1,0 @@
-export * from './Canvas'
-export * from './events'
-export * from '../'
